created models submodule
