+++ mlcvs/models/__init__.py
@@ -1,0 +1,2 @@
+from .linear import LinearCV
+from .nn import NeuralNetworkCV
