+++ mlcvs/models/nn.py
@@ -1,308 +1,15 @@
-"""Collective variables base models."""
+"""Neural network base class."""
 
-__all__ = ["LinearCV", "NeuralNetworkCV"]
+__all__ = ["NeuralNetworkCV"]
 
 import torch
 import numpy as np
 import pandas as pd
 from warnings import warn
-from . import optim
 
-# util for computing mean and range over dataset
+from .. import optim
+from .utils import normalize,compute_mean_range
 
-def compute_mean_range(x: torch.Tensor, print_values=False):
-    """
-    Compute mean and range of values over dataset (for inputs / outputs standardization)
-    """
-    Max, _ = torch.max(x, dim=0)
-    Min, _ = torch.min(x, dim=0)
-
-    Mean = (Max + Min) / 2.0
-    Range = (Max - Min) / 2.0
-
-    if print_values:
-        print("Standardization enabled.")
-        print("Mean:", Mean.shape, "-->", Mean)
-        print("Range:", Range.shape, "-->", Range)
-    if (Range < 1e-6).nonzero().sum() > 0:
-        print(
-            "[Warninthe follig] Normalization: owing features have a range of values < 1e-6:",
-            (Range < 1e-6).nonzero(),
-        )
-        Range[Range < 1e-6] = 1.0
-
-    return Mean, Range
-
-def normalize(
-    x: torch.Tensor, Mean: torch.Tensor, Range: torch.Tensor
-) -> (torch.Tensor):
-    """
-    Compute standardized inputs/outputs (internal).
-
-    Parameters
-    ----------
-    x: torch.Tensor
-        input/output
-    Mean: torch.Tensor
-        mean values to be subtracted.
-    Range: torch.Tensor
-        interval range to be divided by.
-
-    Returns
-    -------
-    out : torch.Tensor
-        standardized inputs/outputs
-    """
-
-    # if shape ==
-
-    if x.ndim == 2:
-        batch_size = x.size(0)
-        x_size = x.size(1)
-
-        Mean_ = Mean.unsqueeze(0).expand(batch_size, x_size)
-        Range_ = Range.unsqueeze(0).expand(batch_size, x_size)
-    elif x.ndim == 1:
-        Mean_ = Mean
-        Range_ = Range
-    else:
-        raise ValueError(
-            "Input tensor must of shape (n_features) or (n_batch,n_features)."
-        )
-
-    return x.sub(Mean_).div(Range_)
-
-class LinearCV(torch.nn.Module):
-    """
-    Linear CV base class.
-
-    Attributes
-    ----------
-    w : torch.Tensor
-        Weights array of linear model
-    b : torch.Tensor
-        Offset array 
-    n_features : int
-        Number of input features
-    feature_names : list
-        List of input features names
-    device_ : torch.Device
-        Device used for the model
-
-    Methods
-    -------
-    __init__(n_features)
-        Create a linear model.
-    forward(X)
-        Project data along linear model
-    get_params()
-        Return saved parameters
-    set_params(dict)
-        Set parameters via dictionaries
-    set_weights(w)
-        Set coefficients
-    set_offset(b)
-        Set linear model
-    plumed_input()
-        Generate PLUMED input file
-    """
-
-    def __init__(self, n_features, device="auto", **kwargs):
-        super().__init__(**kwargs)
-
-        # Device
-
-        if device == "auto":
-            self.device_ = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-        else:
-            self.device_ = device
-
-        # Initialize parameters
-        self.n_features = n_features
-        weight = torch.eye(n_features).to(device=self.device_)
-        offset = torch.zeros(n_features).to(device=self.device_)
-        self.register_buffer("w", weight)
-        self.register_buffer("b", offset)
-
-        # Generic attributes
-        self.name_ = "LinearCV"
-        self.feature_names = ["x" + str(i) for i in range(n_features)]
-
-        # Input normalization
-        self.normIn = False
-        self.normOut = False
-
-    def forward(self, X):
-        """
-        Project data along linear components.
-
-        Parameters
-        ----------
-        X : array-like of shape (n_samples, n_features) or (n_features)
-            Inference data.
-
-        Returns
-        -------
-        s : array-like of shape (n_samples, n_classes-1)
-            Linear projection of inputs.
-        """
-        if type(X) == pd.DataFrame:
-            X = torch.Tensor(X.values).to(device=self.device_)
-        elif type(X) != torch.Tensor:
-            X = torch.Tensor(X).to(device=self.device_)
-
-        if self.normIn:
-            X = normalize(X, self.MeanIn, self.RangeIn)
-
-        s = torch.matmul(X - self.b, self.w)
-
-        if self.normOut:
-            s = normalize(s, self.MeanOut, self.RangeOut)
-
-        return s
-
-    def set_weights(self, w):
-        """
-        Set weights of linear combination.
-
-        Parameters
-        ----------
-        w : torch.tensor
-            weights
-
-        """
-        self.w = w.to(self.device_)
-
-    def set_offset(self, b):
-        """
-        Set linear offset
-
-        Parameters
-        ----------
-        b : torch.tensor
-            offset
-
-        """
-        self.b = b.to(self.device_)
-
-    def get_params(self):
-        """
-        Return saved parameters.
-
-        Returns
-        -------
-        out : namedtuple
-            Parameters
-        """
-        return vars(self)
-
-    def set_params(self, dict_params):
-        """
-        Set parameters.
-
-        Parameters
-        ----------
-        dict_params : dictionary
-            Parameters
-
-        """
-
-        for key in dict_params.keys():
-            if hasattr(self, key):
-                setattr(self, key, dict_params[key])
-            else:
-                raise AttributeError(
-                    f"{self.__class__.__name__} has no attribute {key}."
-                )
-
-    def plumed_input_combine(self): # TODO REMOVE
-        """
-        Generate PLUMED input file
-
-        Deprecated (todo remove)
-
-        Returns
-        -------
-        out : string
-            PLUMED input file
-        """
-
-        # count number of output components
-        weights = self.w.cpu().numpy()
-        offset = self.b.cpu().numpy()
-        n_cv = 1 if weights.ndim == 1 else weights.shape[1]
-
-        out = ""
-        for i in range(n_cv):
-            if n_cv == 1:
-                out += f"{self.name_}: COMBINE ARG="
-            else:
-                out += f"{self.name_}{i+1}: COMBINE ARG="
-            for j in range(self.n_features):
-                out += f"{self.feature_names[j]},"
-            out = out[:-1]
-
-            out += " COEFFICIENTS="
-            for j in range(self.n_features):
-                out += str(np.round(weights[j, i], 6)) + ","
-            out = out[:-1]
-            if not np.all(offset == 0):
-                out += " PARAMETERS="
-                for j in range(self.n_features):
-                    out += str(np.round(offset[j, i], 6)) + ","
-            out = out[:-1]
-
-            out += " PERIODIC=NO\n"
-            
-        return out
-
-    def plumed_input(self):
-        """
-        Generate PLUMED input file
-
-        Returns
-        -------
-        out : string
-            PLUMED input file
-        """
-        
-        # count number of output components
-        weights = self.w.cpu().numpy()
-        offset = self.b.cpu().numpy()
-        n_cv = 1 if weights.ndim == 1 else weights.shape[1]
-
-        # get normalization
-        if self.normIn:
-            mean = self.MeanIn.cpu().numpy()
-
-        out = ""
-        for i in range(n_cv):
-            if n_cv == 1:
-                out += f"{self.name_}: CUSTOM ARG="
-            else:
-                out += f"{self.name_}{i+1}: CUSTOM ARG="
-            for j in range(self.n_features):
-                out += f"{self.feature_names[j]},"
-            out = out[:-1]
-
-            out += " VAR="
-            for j in range(self.n_features):
-                out += f"x{j},"
-            out = out[:-1]
-
-            out += " FUNC="
-            for j in range(self.n_features):
-                w = weights[j, i]
-                s = "+" if w > 0 else ""
-                if self.normIn:
-                    m = mean[j]
-                    s2 = "+" if m < 0 else "-"
-                    out += f'{s}{w:.6f}*(x{j}{s2}{np.abs(m):.6f})'
-                else:
-                    out += f"{s}{w:.6f}*x{j}"
-
-            out += " PERIODIC=NO\n"
-        return out
 
 class NeuralNetworkCV(torch.nn.Module):
     """
